# Conflicts:
#	le-db/src/main/scripts/db_upgrade/pls_multitenant/mysql.sql

+++ le-admin/src/main/java/com/latticeengines/admin/service/impl/AdminRecycleTenantJobCallable.java
@@ -3,6 +3,7 @@ package com.latticeengines.admin.service.impl;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.concurrent.Callable;
@@ -15,6 +16,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.latticeengines.auth.exposed.entitymanager.GlobalAuthUserTenantRightEntityMgr;
+import com.latticeengines.baton.exposed.service.BatonService;
+import com.latticeengines.domain.exposed.admin.LatticeProduct;
 import com.latticeengines.domain.exposed.auth.GlobalAuthUser;
 import com.latticeengines.domain.exposed.auth.GlobalAuthUserTenantRight;
 import com.latticeengines.domain.exposed.camille.CustomerSpace;
@@ -44,6 +47,7 @@ public class AdminRecycleTenantJobCallable implements Callable<Boolean> {
     private AdminProxy adminProxy;
     private com.latticeengines.security.exposed.service.TenantService tenantService;
     private GlobalAuthUserTenantRightEntityMgr userTenantRightEntityMgr;
+    private BatonService batonService;
 
     private static final Logger log = LoggerFactory.getLogger(AdminRecycleTenantJobCallable.class);
 
@@ -54,6 +58,7 @@ public class AdminRecycleTenantJobCallable implements Callable<Boolean> {
         this.adminProxy = builder.adminProxy;
         this.tenantService = builder.tenantService;
         this.userTenantRightEntityMgr = builder.userTenantRightEntityMgr;
+        this.batonService = builder.batonService;
     }
 
     @Override
@@ -62,14 +67,20 @@ public class AdminRecycleTenantJobCallable implements Callable<Boolean> {
         List<Tenant> tempTenants = tenantService.getTenantByTypes(Arrays.asList(TenantType.POC, TenantType.STAGING));
         if (CollectionUtils.isNotEmpty(tempTenants)) {
             log.info("Tenants size is " + tempTenants.size());
+            Set<String> tenantsWithInvalidSettings = new HashSet<>();
             for (Tenant tenant : tempTenants) {
                 log.info("begin dealing with tenant " + tenant.getName());
                 if (tenant.getExpiredTime() == null) {
                     continue;
                 }
+                CustomerSpace space = CustomerSpace.parse(tenant.getId());
+                if (batonService.hasProduct(space, LatticeProduct.DCP)) {
+                    log.info("skip recycling the dnb connect tenant {}", space);
+                    continue;
+                }
                 long expiredTime = tenant.getExpiredTime();
                 long currentTime = LocalDateTime.now().atZone(ZoneId.systemDefault()).toInstant().toEpochMilli();
-                CustomerSpace space = CustomerSpace.parse(tenant.getId());
+
                 // send email two weeks before user can't access tenant
                 if (expiredTime - EMAIL_PERIOD < currentTime && currentTime < expiredTime) {
                     int days = (int) Math.ceil((expiredTime - currentTime) / TimeUnit.DAYS.toMillis(1));
@@ -101,10 +112,19 @@ public class AdminRecycleTenantJobCallable implements Callable<Boolean> {
                     });
 
                 } else if (currentTime > expiredTime + INACTIVE_PERIOD) {
+                    if (!TenantStatus.INACTIVE.equals(tenant.getStatus())) {
+                        log.info("tenant status is not inactive won't be deleted for {}", space);
+                        tenantsWithInvalidSettings.add(space.toString());
+                        continue;
+                    }
                     adminProxy.deleteTenant(space.getContractId(), space.getTenantId());
                     log.info(String.format("tenant %s has been deleted", tenant.getName()));
                 }
 
+            }
+            if (CollectionUtils.isNotEmpty(tenantsWithInvalidSettings)) {
+                throw new RuntimeException(String.format("unsupported settings for tenant %s",
+                        tenantsWithInvalidSettings));
             }
         }
 
@@ -167,6 +187,7 @@ public class AdminRecycleTenantJobCallable implements Callable<Boolean> {
         private AdminProxy adminProxy;
         private com.latticeengines.security.exposed.service.TenantService tenantService;
         private GlobalAuthUserTenantRightEntityMgr userTenantRightEntityMgr;
+        private BatonService batonService;
 
         public Builder() {
 
@@ -201,5 +222,11 @@ public class AdminRecycleTenantJobCallable implements Callable<Boolean> {
             this.userTenantRightEntityMgr = userTenantRightEntityMgr;
             return this;
         }
+
+        public Builder batonService(BatonService batonService) {
+            this.batonService = batonService;
+            return this;
+        }
+
     }
 }
